Update files from daily changes before hosting project on Azure

+++ src/main/frontend/src/components/ListUserComponent.jsx
@@ -21,6 +21,12 @@ export default class ListUserComponent extends Component {
 
   /*Add component of UserService to fetch users from DB (REST API calls come here) */
   componentDidMount() {
+    {
+      /* IMPORTANT !!!!!!!!!!!!!! add page load event if button is clicked and
+              user that is loged in is not admin only user that is 
+              logged in data is displayed and not all of the users */
+    }
+
     UserService.getUsers().then((res) => {
       this.setState({ users: res.data });
     });
@@ -56,7 +62,6 @@ export default class ListUserComponent extends Component {
             startIcon={<AddBoxIcon />}
             onClick={this.addUser}
           >
-            {" "}
             Add User{" "}
           </Button>
         </div>
@@ -102,7 +107,12 @@ export default class ListUserComponent extends Component {
                         variant="contained"
                         startIcon={<DeleteIcon />}
                         onClick={() => {
-                          this.deleteUser(user.id);
+                          if (
+                            window.confirm(
+                              "Are you sure you wish to delete this user?"
+                            )
+                          )
+                            this.deleteUser(user.id);
                         }}
                         classname="p-5"
                       >
